requirements.txt
minio 
python-dotenv
uvicorn